requirements.txt
aiodns==3.0.0
aiogram==2.20
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
alembic==1.7.7
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
environs==9.5.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
multidict==6.0.2
orjson==3.6.8
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyaes==1.6.1
pycares==4.1.2
pycparser==2.21
pyparsing==3.0.8
Pyrogram==2.0.19
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
qrcode==7.3.1
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a22
TgCrypto==1.2.3
typing_extensions==4.2.0
ujson==5.2.0
yarl==1.7.2
